% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_branch_module.R
\name{build_branch_module}
\alias{build_branch_module}
\title{CALLED BY build_levels_based_clique() to build cluster for the two branches of the level x}
\usage{
build_branch_module(
  dfClust,
  actuC,
  newC,
  scm,
  scm_nm,
  clStop,
  tfs,
  setGenesHC,
  goHC,
  gda,
  cliqueTargets,
  adjMethod,
  cutoff,
  qtarget,
  qprobInt,
  posRatio
)
}
\arguments{
\item{dfClust}{dataframe as Cell | Cluster}

\item{actuC}{number of actual cluster}

\item{newC}{integer used to define the next cluster}

\item{scm}{single cell RNA-seq matrix}

\item{scm_nm}{single cell matrix not normalized/filtered}

\item{clStop}{vector of indices to determine when to stop}

\item{tfs}{vector of TFs names or IDs}

\item{setGenesHC}{list of set of genes used for each splitting}

\item{goHC}{list of GO for each set of genes used for each splitting}

\item{gda}{data.frame of GO slim annotations}

\item{cliqueTargets}{list of each antagonistic clique & targets used for the splitting}

\item{adjMethod}{multiple test correction to perform}

\item{cutoff}{p-adjusted p-value cutoff used for the enrichment}
}
\value{
list(dfClust,setGenesHC,goHC,cliqueTargets)
}
\description{
CALLED BY build_levels_based_clique() to build cluster for the two branches of the level x
}
\author{
Celine Barlier
}
