% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_levels_based_clique.R
\name{build_levels_based_clique}
\alias{build_levels_based_clique}
\title{RECURSIVE FUNCTION: Build the child level: split in two the parent}
\usage{
build_levels_based_clique(
  dfClust,
  scm,
  scm_nm,
  firstLevel,
  clStop,
  lev = 2,
  tfs,
  setGenesHC,
  goHC,
  gda,
  cliqueTargets,
  adjMethod,
  cutoff,
  qtarget,
  qprobInt,
  posRatio
)
}
\arguments{
\item{dfClust}{dataframe used to build the final object}

\item{scm}{single cell matrix normalized}

\item{scm_nm}{single cell matrix not normalized/filtered}

\item{firstLevel}{True if level 1, F if level > 1}

\item{clStop}{vector of indices to determine when to stop}

\item{lev}{level number tracking}

\item{tfs}{vector of TFs names or IDs}

\item{setGenesHC}{list of set of genes used for each splitting}

\item{goHC}{list of GO for each set of genes used for each splitting}

\item{gda}{data.frame of GO slim annotations}

\item{cliqueTargets}{list of each antagonistic clique & targets used for the splitting}

\item{adjMethod}{multiple test correction to perform}

\item{cutoff}{p-adjusted p-value cutoff used for the enrichment}
}
\value{
list(dfClust,setGenesHC,goHC,cliqueTargets)
}
\description{
RECURSIVE FUNCTION: Build the child level: split in two the parent
}
\author{
Celine Barlier
}
