% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_object.R
\name{build_object}
\alias{build_object}
\title{Build the functional splitting object}
\usage{
build_object(t, scm)
}
\arguments{
\item{t}{list of results (cliques,GO,clusters)}

\item{scm}{single cell matrix}
}
\value{
functional_split object
}
\description{
Build the functional splitting object
}
\author{
Celine Barlier
}
