% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareExp.R
\name{compareExp}
\alias{compareExp}
\title{Identify if target is expressed in similar cells as regulator and not expressed in antagonistic TFs cells}
\usage{
compareExp(mtx, tf, candidatesTargets)
}
\arguments{
\item{mtx}{single cell matrix}

\item{tf}{regulator name}

\item{candidatesTargets}{vector of potential target genes}
}
\value{
candidateTargets vector order by the genes fullfiling all the criteria (best target -> worst target)
}
\description{
Identify if target is expressed in similar cells as regulator and not expressed in antagonistic TFs cells
}
\author{
Celine Barlier
}
