% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillEnrichFunPartObj.R
\name{fillEnrichFunPartObj}
\alias{fillEnrichFunPartObj}
\title{Update FunPart object structure
Add modules & functionalenrich slots to provide more details}
\usage{
fillEnrichFunPartObj(FunPartObj, m, gda, padjMeth = adjMethod, padj = cutoff)
}
\arguments{
\item{FunPartObj}{FunPart object}

\item{m}{Initial single cell matrix (cells in columns, genes in rows)}

\item{gda}{Enrichment ressource file}

\item{padjMeth}{P-adjusted correction method}

\item{padj}{p-adjusted value cutoff}
}
\value{
FunPart object updated slots
}
\description{
Update FunPart object structure
Add modules & functionalenrich slots to provide more details
}
\author{
Celine Barlier
}
