% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichedBP.R
\name{getEnrichedBP}
\alias{getEnrichedBP}
\title{GO enrichment of a set of genes}
\usage{
getEnrichedBP(dataset, setGenes, gda, adjMethod, cutoff)
}
\arguments{
\item{dataset}{single-cell dataset (cells in columns / genes in rows)}

\item{setGenes}{vector of genes (from one antagonistic module)}

\item{gda}{data.frame of GO slim annotations}

\item{adjMethod}{multiple test correction to perform}

\item{cutoff}{p-adjusted p-value cutoff used for the enrichment}
}
\value{
data.frame of enriched BP
}
\description{
GO enrichment of a set of genes
}
\author{
Celine Barlier
}
