% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichedBPManual.R
\name{getEnrichedBPManual}
\alias{getEnrichedBPManual}
\title{Get all the BPs enriched (used to fill in the new structure of FunPart object)}
\usage{
getEnrichedBPManual(
  dataset,
  setGenes,
  gda,
  padjMeth = adjMethod,
  padj = cutoff
)
}
\arguments{
\item{dataset}{Initial single cell matrix (cells in columns, genes in rows)}

\item{setGenes}{Enrichment ressource file}

\item{gda}{Enrichment ressource file}

\item{padjMeth}{P-adjusted correction method}

\item{padj}{p-adjusted value cutoff}
}
\value{
vector of functional enrichment
}
\description{
Get all the BPs enriched (used to fill in the new structure of FunPart object)
}
\author{
Celine Barlier
}
