% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanRatioExpModule.R
\name{getMeanRatioExpModule}
\alias{getMeanRatioExpModule}
\title{Get mean ratio cell expressing TFs of the module}
\usage{
getMeanRatioExpModule(m, moduletfs, cells)
}
\arguments{
\item{m}{normalized/cleaned single cell expression matrix}

\item{moduletfs}{vector of TFs names belonging to the same module}

\item{cells}{cells to use to calculate the expression ratio}
}
\value{
mean ratio of cells expressing moduletfs
}
\description{
Get mean ratio cell expressing TFs of the module
}
\author{
Celine Barlier
}
