% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleFunctionalState.R
\name{getModuleFunctionalState}
\alias{getModuleFunctionalState}
\title{Process FunPart objects content into a table with all results}
\usage{
getModuleFunctionalState(funpartobj)
}
\arguments{
\item{funpartobj}{FunPart output object}
}
\value{
dataframe with the formatted results: modules, branch, type, tfs, genes and enrichment
}
\description{
Process FunPart objects content into a table with all results
}
\author{
Celine Barlier
}
