% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clust.R
\name{hierarchical_clust}
\alias{hierarchical_clust}
\title{Hierarchical clustering using Pheatmap package}
\usage{
hierarchical_clust(mhc)
}
\arguments{
\item{mhc}{single cell RNA-seq with genes being the ones of the best antagonistic and functionally enriched modules}
}
\value{
dendrogram object
}
\description{
Hierarchical clustering using Pheatmap package
}
\author{
Celine Barlier
}
