% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_obj.R
\name{init_obj}
\alias{init_obj}
\title{Initialise functional splitting object elements}
\usage{
init_obj(cl, top_set_genes, best_set_genes_go, cliqueRes)
}
\arguments{
\item{cl}{vector of clusters}

\item{top_set_genes}{set of genes modules used}

\item{best_set_genes_go}{details about the genes used}

\item{cliqueRes}{}
}
\value{
list of functional_split object elements init
}
\description{
Initialise functional splitting object elements
}
\author{
Celine Barlier
}
