% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_functional_splitting.R
\name{run_functional_splitting}
\alias{run_functional_splitting}
\title{MAIN FUNCTION: functional splitting}
\usage{
run_functional_splitting(
  scm,
  tfs,
  gda,
  norm = T,
  qtarget = 0.9,
  adjMethod = "BH",
  cutoff = 0.05,
  percExp = 10,
  qExp = 0.25,
  qprobInt = c(0.025, 0.975),
  posRatio = 1
)
}
\arguments{
\item{scm}{single cell matrix (cells in columns / genes in rows)}

\item{tfs}{vector of TFs names / ID depending on the genes in the scm}

\item{gda}{BP GO file used for the enrichment analysis}

\item{norm}{T = perform seurat normalization (default), F = do not perform the normalization}

\item{adjMethod}{multiple test correction to perform, either BH (default) or bonferroni}

\item{cutoff}{p-adjusted value cutoff for the functional enrichment}

\item{percExp}{percentage of cells a gene needs to be expressed in to be considered}

\item{qExp}{quantile of gene expression that will be considered as too lowly expressed and so removed}

\item{qprobInt}{quantile of interactions to consider as strong and to use for clique identification}

\item{posRatio}{positive ratio strenght inside a clique}

\item{q}{top X target genes to consider for each identified cliques (default = top 1 percent)}
}
\value{
functional_split object
}
\description{
MAIN FUNCTION: functional splitting
}
\author{
Celine Barlier
}
