% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGOanalysis_listSetGenes.R
\name{runGOanalysis_listSetGenes}
\alias{runGOanalysis_listSetGenes}
\title{Run BP enrichment for a list of set of genes}
\usage{
runGOanalysis_listSetGenes(dataset, listSetGenes, gda, adjMethod, cutoff)
}
\arguments{
\item{dataset}{single-cell dataset (cells in columns / genes in rows)}

\item{listSetGenes}{list of identified antagonistic modules}

\item{gda}{data.frame of GO slim annotations}

\item{adjMethod}{multiple test correction to perform}

\item{cutoff}{p-adjusted p-value cutoff used for the enrichment}
}
\value{
df of cliques + GO + FC enrichment
}
\description{
Run BP enrichment for a list of set of genes
}
\author{
Celine Barlier
}
