% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_new_clique_from_list.R
\name{test_new_clique_from_list}
\alias{test_new_clique_from_list}
\title{Test different cliques to be used if one get drop out all along the criteria filters}
\usage{
test_new_clique_from_list(scm, best_set_genes_go, incS = 2, cellNum = 5)
}
\arguments{
\item{best_set_genes_go}{details about the genes used}

\item{incS}{clique number to start the search with}

\item{cellNum}{minimum number of cells to be present in a cluster to be considered}

\item{cl}{vector of clusters}

\item{top_set_genes}{set of genes modules used}
}
\value{
list of functional_split object elements init
}
\description{
Test different cliques to be used if one get drop out all along the criteria filters
}
\author{
Celine Barlier
}
