% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareClk.R
\name{compareClk}
\alias{compareClk}
\title{Identify if a clique is unique}
\usage{
compareClk(v, clk)
}
\arguments{
\item{v}{vectors of TFs}

\item{clk}{liste of positive cliques}
}
\value{
true or false
}
\description{
Identify if a clique is unique
}
\author{
Celine Barlier
}
