% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_norm_data.R
\name{filter_norm_data}
\alias{filter_norm_data}
\title{Clean (and normalize) the data}
\usage{
filter_norm_data(mtx, norm, percExp, qExp)
}
\arguments{
\item{mtx}{single cell matrix (cells in columns / genes in rows)}

\item{norm}{perform Seurat's normalization (True) or not (False)}

\item{percExp}{percentage of cells in which the gene has to be expressed to be considered}

\item{qExp}{quantile above which the genes are considered (QC step)}
}
\value{
cleaned (and normalized) matrix
}
\description{
Clean (and normalize) the data
}
\author{
Celine Barlier
}
